refactor: move handle image picker function to ImageHelper.java for further usage
