+++ frontend/app/src/main/AndroidManifest.xml
@@ -44,7 +44,7 @@
             android:name=".authenticate.register.SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
