+++ frontend/app/src/main/java/com/example/gathernow/authenticate/UserLocalDataSource.java
@@ -3,14 +3,14 @@ package com.example.gathernow.authenticate;
 import android.content.Context;
 import android.content.SharedPreferences;
 
-public class UserDataSource {
+public class UserLocalDataSource {
 
     private final SharedPreferences sharedPreferences;
     private final SharedPreferences.Editor editor;
     private static final String PREF_NAME = "UserSession";
     private static final String KEY_USER_ID = "user_id";
 
-    public UserDataSource(Context context) {
+    public UserLocalDataSource(Context context) {
         sharedPreferences = context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
         editor = sharedPreferences.edit();
     }
